Can't exit 1 if jshstub is sourced!

+++ code/shellscript/jshstub/jshstub.sh
@@ -21,10 +21,12 @@ SCRIPTNAME=`basename "$SCRIPTFILE"`
 	# exit 1
 # fi
 
+OKTOGO=true
+
 if test "$SCRIPTNAME" = jshstub
 then
 	echo "jshstub: Refusing to retrieve another copy of jshstub!" >&2
-	exit 1
+	OKTOGO=
 fi
 
 if test ! -L "$SCRIPTFILE"
@@ -35,23 +37,31 @@ then
 		SCRIPTFILE="$JPATH/tools/$SCRIPTFILE"
 	else
 		echo "jshstub: Aborting because $SCRIPTFILE is not a symlink!" >&2
-		exit 1
+		OKTOGO=
 	fi
 fi
 
-rm -f "$SCRIPTFILE"
-
-wget "http://hwi.ath.cx/jshstubtools/$SCRIPTNAME" -O "$SCRIPTFILE" > /tmp/jshstub_wget.log
-
-if test ! "$?" = 0
+if test "$OKTOGO"
 then
-	echo "jshstub: Error: failed to retrieve http://hwi.ath.cx/jshstubtools/$SCRIPTNAME" >&2
-	ln -s "$JPATH/tools/jshstub" "$SCRIPTFILE"
-	exit 1
+
+	rm -f "$SCRIPTFILE"
+
+	wget "http://hwi.ath.cx/jshstubtools/$SCRIPTNAME" -O "$SCRIPTFILE" > /tmp/jshstub_wget.log
+
+	if test ! "$?" = 0
+	then
+
+		echo "jshstub: Error: failed to retrieve http://hwi.ath.cx/jshstubtools/$SCRIPTNAME" >&2
+		echo "jshstub: Replacing removed symlink"
+		ln -s "$JPATH/tools/jshstub" "$SCRIPTFILE"
+
+	else
+
+		echo "jshstub: got script \"$SCRIPTNAME\" ok, now running $SCRIPTNAME $* ..." >&2
+
+		chmod a+x "$SCRIPTFILE"
+		"$SCRIPTFILE" "$@"
+
+	fi
+
 fi
-
-chmod a+x "$SCRIPTFILE"
-
-echo "jshstub: got script \"$SCRIPTNAME\" ok, now running $SCRIPTNAME $* ..." >&2
-
-"$SCRIPTFILE" "$@"
